Refactor: Better structure to folders and files
