+++ src/Layout/Footer/Footer.module.scss
@@ -1,4 +1,4 @@
-@use '../../../styles/variables.scss' as *;
+@use '../../styles/variables.scss' as *;
 
 .footer {
   margin-top: 1rem;
