+++ src/Layout/Header/Header.module.scss
@@ -1,4 +1,4 @@
-@use './../../../styles/variables.scss' as *;
+@use '../../styles/variables.scss' as *;
 
 .header {
   position: relative;
