+++ src/Layout/Header/HeaderActions/HeaderActions.module.scss
@@ -1,5 +1,5 @@
-@use './../../../../styles/variables.scss' as *;
-@use '../../../../styles/mixins/layout.scss' as *;
+@use '../../../styles/variables.scss' as *;
+@use '../../../styles/mixins/layout.scss' as *;
 
 .actionsRow {
   @include rowWrapper;
