+++ src/Layout/Header/HeaderMenu/HeaderMenu.module.scss
@@ -1,5 +1,5 @@
-@use '../../../../styles/variables.scss' as *;
-@use '../../../../styles/mixins/layout.scss' as *;
+@use '../../../styles/variables.scss' as *;
+@use '../../../styles/mixins/layout.scss' as *;
 
 .menuRow {
   @include rowWrapper;
