+++ src/features/Cart/AddToCartButton/AddToCartButton.module.scss
@@ -1,4 +1,4 @@
-@use '../../styles/variables.scss' as *;
+@use '../../../styles/variables.scss' as *;
 
 .icon {
   width: 1.8rem;
