+++ src/features/Cart/RemoveFromCart/RemoveFromCartButton.jsx
@@ -1,5 +1,5 @@
 import { IoIosClose } from 'react-icons/io';
-import IconButton from '../IconButton/IconButton';
+import IconButton from '../../../components/IconButton/IconButton';
 import styles from './RemoveFromCartButton.module.scss';
 
 export default function RemoveFromCartButton({ onClick }) {
