+++ src/features/Cart/RemoveFromCart/RemoveFromCartButton.module.scss
