+++ src/features/Product/Expandable.module.scss
